perfecto_labs/requirements.txt
customtkinter>=5.2.0
mysql-connector-python>=8.2.0
Pillow>=10.0.0
fpdf2>=2.7.0
bcrypt>=4.1.0
python-escpos>=3.0
